provide helper script to finish setup of the container
When starting the container the user has to do some manual steps in
order to be ready to run puppet. This helper script will take care
of those steps, so that the user only has to run setup.sh.

To not convolute the root directory the script was purposefully
added in the bin subdirectory, where we can add further scripts in
the future, e.g. helper scripts to run puppet without having to
supply all the necessary arguments each time.

For convenience the bin directory is added to $PATH.