ensure container is ready on login
The previous way of running the bin/setup.sh script didn't work
quite correct. Also, it still required manual interaction. This
change will automatically ensure apt is updated and that SSH keys
are loaded on login to the container, so the user can start right
away with running r10k or puppet.